+++ src/Class/class.controller.ts
@@ -1,9 +1,19 @@
 import { NextFunction, Request, Response } from 'express';
 import * as classServices from './class.services'
 import createHttpError from 'http-errors';
-
-async function createClass(req: Request, res: Response, next: NextFunction): Promise<void> {
+import { Role } from '@prisma/client';
+interface AuthenticatedRequest extends Request {
+    payload: {
+        role: Role;
+    };
+}
+async function createClass(req: AuthenticatedRequest, res: Response, next: NextFunction) {
     try {
+        if (req.payload.role !== Role.ADMIN) {
+            return res.status(403).json({ success: false, message: 'Forbidden: You are not authorized to access this operation' });
+        }
+
+
         const { name } = req.body
 
         const newClass = await classServices.addClass(name)
@@ -17,11 +27,15 @@ async function createClass(req: Request, res: Response, next: NextFunction): Pro
         }
 
     } catch (error) {
-        next(createHttpError(error.statusCode, error.message))
+        next(createHttpError(error.statusCode || 500, error.message))
     }
 }
-async function updateClass(req: Request, res: Response, next: NextFunction): Promise<void> {
+async function updateClass(req: AuthenticatedRequest, res: Response, next: NextFunction) {
     try {
+        if (req.payload.role !== Role.ADMIN) {
+            return res.status(403).json({ success: false, message: 'Forbidden: You are not authorized to access this operation' });
+        }
+
         const id = parseInt(req.params.id)
 
         const { name } = req.body
@@ -37,11 +51,15 @@ async function updateClass(req: Request, res: Response, next: NextFunction): Pro
         }
 
     } catch (error) {
-        next(createHttpError(error.statusCode, error.message))
+        next(createHttpError(error.statusCode || 500, error.message))
     }
 }
-async function removeClass(req: Request, res: Response, next: NextFunction): Promise<void> {
+async function removeClass(req: AuthenticatedRequest, res: Response, next: NextFunction) {
     try {
+        if (req.payload.role !== Role.ADMIN) {
+            return res.status(403).json({ success: false, message: 'Forbidden: You are not authorized to access this operation' });
+        }
+
         const id = parseInt(req.params.id)
 
         const deleteClass = await classServices.deleteClass(id)
@@ -55,12 +73,15 @@ async function removeClass(req: Request, res: Response, next: NextFunction): Pro
         }
 
     } catch (error) {
-        next(createHttpError(error.statusCode, error.message))
+        next(createHttpError(error.statusCode || 500, error.message))
     }
 }
 
-async function getAllClass(req: Request, res: Response, next: NextFunction): Promise<void> {
+async function getAllClass(req: AuthenticatedRequest, res: Response, next: NextFunction) {
     try {
+        if (req.payload.role !== Role.ADMIN) {
+            return res.status(403).json({ success: false, message: 'Forbidden: You are not authorized to access this operation' });
+        }
 
         const allClass = await classServices.getAllClasses()
 
@@ -71,13 +92,33 @@ async function getAllClass(req: Request, res: Response, next: NextFunction): Pro
                 data: allClass
             })
         }
-        
+
     } catch (error) {
-        next(createHttpError(error.statusCode, error.message))
+        next(createHttpError(error.statusCode || 500, error.message))
+    }
+}
+async function getOneClass(req: AuthenticatedRequest, res: Response, next: NextFunction) {
+    try {
+        if (req.payload.role !== Role.ADMIN) {
+            return res.status(403).json({ success: false, message: 'Forbidden: You are not authorized to access this operation' });
+        }
+        const id = parseInt(req.params.id)
+        const allClass = await classServices.getOneClass(id)
+
+        if (allClass) {
+            res.status(200).json({
+                success: true,
+                message: "Class Fetched",
+                data: allClass
+            })
+        }
+
+    } catch (error) {
+        next(createHttpError(error.statusCode || 500, error.message))
     }
 }
 
 
 
 
-export { createClass, getAllClass, updateClass, removeClass }
+export { createClass, getAllClass, updateClass, removeClass,getOneClass }
